Type uint64 amounts in trade-record/offer collections; fix offered/requested string typing
The int->uint64 BigInt audit's grep missed int used inside Record/array/tuple
positions. Fix the amount-bearing ones: TradeRecord.pending and
create_offer_for_ids request offer values -> uint64 (number | bigint).
Also correct TradeRecord summary.offered/requested to Record<str, str>: those
offer amounts are stringified on the wire since chia 2.6.0 (a string-vs-number
fix, separate from BigInt).

+++ src/api/rpc/wallet/index.ts
@@ -1721,7 +1721,7 @@ export async function cat_get_asset_id<T extends TRPCAgent | TDaemon>(
 export const create_offer_for_ids_command = "create_offer_for_ids";
 export type create_offer_for_ids_command = typeof create_offer_for_ids_command;
 export type TCreateOfferForIdsRequest = {
-  offer: Record<str, str | int>; // values are canonically strings as of chia-blockchain 2.6.0
+  offer: Record<str, str | uint64>; // values are canonically strings as of chia-blockchain 2.6.0
 
   fee?: uint64;
   validate_only?: bool;
